+++ package.json
@@ -17,15 +17,15 @@
     "custom_typings/"
   ],
   "scripts": {
-    "clean": "mkdir -p lib/ && rm -rf lib/*",
+    "clean": "mkdir -p lib/ && rm -rf lib/",
     "compile": "tsc",
-    "build": "npm run compile",
+    "build": "mkdir -p lib/ && cp -r src/test lib/ && npm run compile",
     "release": "npm run compile",
     "build:watch": "watch 'npm run build' ./src",
     "lint": "gulp lint",
     "test": "npm run clean && npm run build && npm run lint && mocha",
     "quicktest": "npm run build && mocha",
-    "test:watch": "watchy -w ./ -i '/(.git|lib|node_modules)/' -- npm run quicktest",
+    "test:watch": "watchy -w src/ -- npm run quicktest",
     "format": "find src test | grep '\\.js$\\|\\.ts$' | xargs ./node_modules/.bin/clang-format --style=file -i"
   },
   "devDependencies": {
