+++ src/javascript/javascript-entity-finder.ts
@@ -14,10 +14,10 @@
 
 import {Program} from 'estree';
 
-import {Visitor} from './estree-visitor';
 import {Descriptor} from '../ast/ast';
 import {EntityFinder} from '../entity/entity-finder';
 
+import {Visitor} from './estree-visitor';
 import {JavaScriptDocument} from './javascript-document';
 
 export interface JavaScriptEntityFinder extends
