+++ src/javascript/javascript-parser.ts
@@ -17,9 +17,9 @@ import * as estraverse from 'estraverse';
 import {Program} from 'estree';
 
 import {Analyzer} from '../analyzer';
-import {Visitor} from './estree-visitor';
 import {Parser} from '../parser/parser';
 
+import {Visitor} from './estree-visitor';
 import {JavaScriptDocument} from './javascript-document';
 
 export class JavaScriptParser implements Parser<JavaScriptDocument> {
