+++ src/polymer/analyze-properties.ts
@@ -14,9 +14,9 @@
 
 import * as estree from 'estree';
 
-import * as esutil from '../javascript/esutil';
-import * as astValue from '../javascript/ast-value';
 import {PropertyDescriptor} from '../ast/ast';
+import * as astValue from '../javascript/ast-value';
+import * as esutil from '../javascript/esutil';
 
 export function analyzeProperties(node: estree.Node) {
   const analyzedProps: PropertyDescriptor[] = [];
