+++ src/test/html/html-import-finder_test.js
@@ -1,0 +1,60 @@
+/**
+ * @license
+ * Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+'use strict';
+
+const assert = require('chai').assert;
+const fs = require('fs');
+const parse5 = require('parse5');
+const path = require('path');
+
+const HtmlDocument = require('../../html/html-document').HtmlDocument;
+const HtmlImportFinder =
+    require('../../html/html-import-finder').HtmlImportFinder;
+
+suite('HtmlImportFinder', () => {
+
+  suite('findImports()', () => {
+    let finder;
+
+    setup(() => {
+      finder = new HtmlImportFinder();
+    });
+
+    test('finds HTML Imports', () => {
+      let contents = `<html><head>
+          <link rel="import" href="polymer.html">
+          <link rel="import" type="css" href="polymer.css">
+          <script src="foo.js"></script>
+          <link rel="stylesheet" href="foo.css"></link>
+        </head></html>`;
+      let ast = parse5.parse(contents);
+      let document = new HtmlDocument({
+        url: 'test.html',
+        contents,
+        ast,
+      });
+      let visit = (visitor) => document.visit([visitor]);
+
+      return finder.findEntities(document, visit).then((entities) => {
+        assert.equal(entities.length, 1);
+        assert.equal(entities[0].type, 'html-import');
+        assert.equal(entities[0].url, 'polymer.html');
+      });
+
+    });
+
+  });
+
+});
