+++ src/test/static/dependencies/inline-and-imports.html
