+++ src/test/static/dependencies/inline-only.html
