+++ src/test/static/dependencies/subfolder/subfolder-sibling.html
