+++ src/test/static/es6-support-simple.js
