+++ src/test/static/es6-support.js
