+++ src/test/static/xhr-document.html
