+++ src/test/static/xhr-text.txt
